requirements/dev.txt
-e file:.
-e file:./source/idea/idea-data-model/src
-e file:./source/idea/library/src
aiofiles==24.1.0
alembic==1.14.0
arrow==1.3.0
assertpy==1.1
astroid==2.15.8
attrs==24.2.0
aws-cdk-asset-awscli-v1==2.2.213
aws-cdk-asset-kubectl-v20==2.1.3
aws-cdk-asset-node-proxy-agent-v6==2.1.0
aws-cdk-aws-batch-alpha==2.72.1a0
aws-cdk-cloud-assembly-schema==38.0.1
aws-cdk-lib==2.171.0
babel==2.16.0
banal==1.0.6
black==24.10.0
blinker==1.9.0
boto3==1.34.155
boto3-stubs-lite[cloudformation,essential]==1.35.68
botocore==1.34.162
botocore-stubs==1.35.68
build==1.2.2.post1
cacheout==0.16.0
cachetools==5.5.0
cattrs==24.1.2
cdk-bootstrapless-synthesizer==2.3.2
cdk-nag==2.34.6
certifi==2024.8.30
cffi==1.17.1
cfn-flip==1.3.0
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
colored==1.4.4
commonmark==0.9.1
constructs==10.4.2
coverage[toml]==7.6.8
cryptography==42.0.8
dataset==1.6.2
defusedxml==0.7.1
dill==0.3.9
distlib==0.3.9
exceptiongroup==1.2.2
execnet==2.1.1
fastcounter==1.1.0
filelock==3.16.1
ghp-import==2.1.0
h11==0.14.0
html5tagger==1.3.0
httptools==0.6.4
idna==3.10
importlib-metadata==8.5.0
importlib-resources==6.4.5
iniconfig==2.0.0
invoke==2.2.0
ipaddress==1.0.23
isort==5.13.2
jinja2==3.1.4
jmespath==1.0.1
jsii==1.105.0
lazy-object-proxy==1.10.0
ldappool==3.0.0
mako==1.3.6
markdown==3.7
markupsafe==3.0.2
mccabe==0.7.0
memory-profiler==0.61.0
mergedeep==1.3.4
mkdocs==1.6.1
mkdocs-get-deps==0.2.0
mkdocs-material==9.5.46
mkdocs-material-extensions==1.3.1
moto==5.0.21
multidict==6.1.0
mypy==0.991
mypy-boto3-cloudformation==1.35.64
mypy-boto3-dynamodb==1.35.60
mypy-boto3-ec2==1.35.67
mypy-boto3-lambda==1.35.68
mypy-boto3-rds==1.35.66
mypy-boto3-s3==1.35.67
mypy-boto3-sqs==1.35.0
mypy-extensions==1.0.0
openapi-schema-pydantic==1.2.4
orjson==3.10.12
outcome==1.3.0.post0
packaging==24.2
paginate==0.5.7
pathspec==0.12.1
pip-tools==7.4.1
platformdirs==4.3.6
pluggy==1.5.0
prettytable==3.12.0
prometheus-client==0.21.0
prompt-toolkit==3.0.48
psutil==5.9.8
publication==0.0.3
pyasn1==0.6.1
pyasn1-modules==0.4.1
pycparser==2.22
pydantic==1.10.19
pyfiglet==1.0.2
pygments==2.18.0
pyhocon==0.3.61
pyjwt==2.9.0
pylint==2.17.7
pymdown-extensions==10.12
pyparsing==2.4.7
pyproject-api==1.8.0
pyproject-hooks==1.2.0
pysocks==1.7.1
pytest==7.2.2
pytest-cov==4.1.0
pytest-mock==3.14.0
pytest-xdist==3.6.1
python-dateutil==2.9.0.post0
python-dynamodb-lock==0.9.1
python-ldap==3.4.4
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
pyyaml==6.0.2
pyyaml-env-tag==0.1
questionary==1.10.0
random-password-generator==2.2.0
regex==2024.11.6
requests==2.31.0
requests-aws4auth==1.3.1
requests-unixsocket==0.3.0
responses==0.25.3
rich==12.6.0
s3transfer==0.10.4
sanic==23.12.2
sanic-ext==23.12.0
sanic-routing==23.12.0
selenium==4.27.0
semver==2.13.0
sh==1.14.3
shortuuid==1.0.13
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlalchemy==1.4.54
supervisor==4.2.5
tomli==2.1.0
tomlkit==0.13.2
tox==4.23.2
tracerite==1.1.1
trio==0.27.0
trio-websocket==0.11.1
troposphere==4.8.3
typeguard==2.13.3
types-awscrt==0.23.0
types-python-dateutil==2.9.0.20241003
types-pyyaml==6.0.12.20240917
types-requests==2.31.0.6
types-s3transfer==0.10.4
types-urllib3==1.26.25.14
types-xmltodict==0.14.0.20241009
typing-extensions==4.12.2
tzdata==2024.2
tzlocal==4.3.1
ujson==5.10.0
urllib3[socks]==1.26.20
uvloop==0.21.0
validators==0.34.0
virtualenv==20.27.1
watchdog==6.0.0
wcwidth==0.2.13
websocket-client==1.8.0
websockets==14.1
werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.0
wsproto==1.2.0
xmltodict==0.14.2
zipp==3.21.0

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
